make logic changes
